+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/PlutusScriptApi.hs
@@ -42,14 +42,19 @@ import Cardano.Ledger.Alonzo.TxInfo (
   runPlutusScript,
   valContext,
  )
-import Cardano.Ledger.Alonzo.TxWits (AlonzoEraTxWits (..), unTxDats)
-import Cardano.Ledger.Alonzo.UTxO (AlonzoScriptsNeeded (..))
-import Cardano.Ledger.BaseTypes (ProtVer (pvMajor), StrictMaybe (..), natVersion)
+import Cardano.Ledger.Alonzo.TxWits (AlonzoEraTxWits (..))
+import Cardano.Ledger.Alonzo.UTxO (
+  AlonzoEraUTxO (getSpendingDatum),
+  AlonzoScriptsNeeded (..),
+  getAlonzoSpendingDatum,
+  getAlonzoSpendingTxIn,
+ )
+import Cardano.Ledger.BaseTypes (ProtVer (pvMajor), natVersion)
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Language (BinaryPlutus (..), Language (..), Plutus (..))
 import Cardano.Ledger.TxIn (TxIn (..))
-import Cardano.Ledger.UTxO (EraUTxO (..), UTxO (..))
+import Cardano.Ledger.UTxO (EraUTxO (..), ScriptsProvided (..), UTxO (..))
 import Cardano.Slotting.EpochInfo (EpochInfo)
 import Cardano.Slotting.Time (SystemStart)
 import Control.Monad (guard)
@@ -70,25 +75,19 @@ import NoThunks.Class (NoThunks)
 
 -- | Only the Spending ScriptPurpose contains TxIn
 getSpendingTxIn :: ScriptPurpose era -> Maybe (TxIn (EraCrypto era))
-getSpendingTxIn = \case
-  Spending txin -> Just txin
-  Minting _policyid -> Nothing
-  Rewarding _rewaccnt -> Nothing
-  Certifying _dcert -> Nothing
+getSpendingTxIn = getAlonzoSpendingTxIn
+{-# DEPRECATED getSpendingTxIn "In favor of `getAlonzoSpendingTxIn`" #-}
 
 -- | Get the Data associated with a ScriptPurpose. Only the Spending
---   ScriptPurpose contains Data. The null list is returned for the other kinds.
+--   ScriptPurpose contains Data. Nothing is returned for the other kinds.
 getDatumAlonzo ::
   (AlonzoEraTxWits era, AlonzoEraTxOut era, EraTx era) =>
   Tx era ->
   UTxO era ->
   ScriptPurpose era ->
   Maybe (Data era)
-getDatumAlonzo tx (UTxO m) sp = do
-  txIn <- getSpendingTxIn sp
-  txOut <- Map.lookup txIn m
-  SJust hash <- Just $ txOut ^. dataHashTxOutL
-  Map.lookup hash (unTxDats $ tx ^. witsTxL . datsTxWitsL)
+getDatumAlonzo tx utxo = getAlonzoSpendingDatum utxo tx
+{-# DEPRECATED getDatumAlonzo "In favor of `getAlonzoSpendingDatum`" #-}
 
 -- ========================================================================
 
@@ -143,10 +142,9 @@ knownToNotBe1Phase scriptsAvailable (sp, sh) = do
 --     if that is not the case, a PredicateFailure is raised in the Utxos rule.
 collectTwoPhaseScriptInputs ::
   forall era.
-  ( EraTx era
-  , MaryEraTxBody era
+  ( MaryEraTxBody era
   , AlonzoEraTxWits era
-  , EraUTxO era
+  , AlonzoEraUTxO era
   , ScriptsNeeded era ~ AlonzoScriptsNeeded era
   , ExtendedUTxO era
   , Script era ~ AlonzoScript era
@@ -168,10 +166,9 @@ collectTwoPhaseScriptInputs ei sysS pp tx utxo =
 
 collectPlutusScriptsWithContext ::
   forall era.
-  ( EraTx era
-  , MaryEraTxBody era
+  ( MaryEraTxBody era
   , AlonzoEraTxWits era
-  , EraUTxO era
+  , AlonzoEraUTxO era
   , ScriptsNeeded era ~ AlonzoScriptsNeeded era
   , ExtendedUTxO era
   , Script era ~ AlonzoScript era
@@ -203,10 +200,10 @@ collectPlutusScriptsWithContext epochInfo sysStart pp tx utxo =
             (Right [])
   where
     costModels = costModelsValid $ pp ^. ppCostModelsL
-    scriptsAvailable = txscripts utxo tx
+    ScriptsProvided scriptsProvided = getScriptsProvided utxo tx
     AlonzoScriptsNeeded scriptsNeeded' = getScriptsNeeded utxo (tx ^. bodyTxL)
     neededAndConfirmedToBePlutus =
-      mapMaybe (knownToNotBe1Phase scriptsAvailable) scriptsNeeded'
+      mapMaybe (knownToNotBe1Phase scriptsProvided) scriptsNeeded'
     getScriptWithRedeemer (sp, script) =
       case indexedRdmrs tx sp of
         Just (d, eu) -> Right (script, sp, d, eu)
@@ -215,7 +212,7 @@ collectPlutusScriptsWithContext epochInfo sysStart pp tx utxo =
       cm <- maybe (Left (NoCostModel lang)) Right $ Map.lookup lang costModels
       case txInfo pp lang epochInfo sysStart utxo tx of
         Right inf ->
-          let datums = maybe id (:) (getDatum tx utxo sp) [d, valContext inf sp]
+          let datums = maybe id (:) (getSpendingDatum utxo tx sp) [d, valContext inf sp]
            in Right $
                 PlutusWithContext
                   { pwcScript = script
@@ -292,8 +289,7 @@ evalPlutusScripts pv tx (plutusWithContext : rest) =
 -- THE SPEC CALLS FOR A SET, BUT THAT NEEDS A BUNCH OF ORD INSTANCES (TxCert)
 -- See additional comments about 'scriptsNeededFromBody' below.
 scriptsNeeded ::
-  ( EraTx era
-  , EraUTxO era
+  ( EraUTxO era
   , ScriptsNeeded era ~ [(ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era))]
   ) =>
   UTxO era ->