+++ eras/conway/impl/src/Cardano/Ledger/Conway/UTxO.hs
@@ -14,10 +14,17 @@ module Cardano.Ledger.Conway.UTxO (
 ) where
 
 import Cardano.Ledger.Alonzo.UTxO (
+  AlonzoEraUTxO (..),
   AlonzoScriptsNeeded (..),
   getAlonzoScriptsHashesNeeded,
   getAlonzoScriptsNeeded,
  )
+import Cardano.Ledger.Babbage.UTxO (
+  getBabbageScriptsProvided,
+  getBabbageSpendingDatum,
+  getBabbageSupplementalDataHashes,
+ )
+
 import Cardano.Ledger.Coin (Coin (Coin))
 import Cardano.Ledger.Conway.Core (
   Era (..),
@@ -33,6 +40,7 @@ import Cardano.Ledger.Conway.Governance.Procedures (
   VotingProcedures (..),
  )
 import Cardano.Ledger.Conway.PParams (ConwayEraPParams, ppDRepDepositL)
+import Cardano.Ledger.Conway.Tx ()
 import Cardano.Ledger.Conway.TxBody (ConwayEraTxBody (..))
 import Cardano.Ledger.Conway.TxCert (
   ConwayEraTxCert,
@@ -125,6 +133,13 @@ instance Crypto c => EraUTxO (ConwayEra c) where
 
   getProducedValue = conwayProducedValue
 
+  getScriptsProvided = getBabbageScriptsProvided
+
   getScriptsNeeded = getConwayScriptsNeeded
 
   getScriptsHashesNeeded = getAlonzoScriptsHashesNeeded -- TODO: This also changes for Conway
+
+instance Crypto c => AlonzoEraUTxO (ConwayEra c) where
+  getSupplementalDataHashes = getBabbageSupplementalDataHashes
+
+  getSpendingDatum = getBabbageSpendingDatum