+++ eras/babbage/impl/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.4.5.1
+## 1.5.0.0
 
-*
+* Switch to using `AlonzoEraUTxO` in rules
+* Deprecate `getDatumBabbage` in favor of `getBabbageSpendingDatum`
+* Add `getBabbageSupplementalDataHashes`
+* Remove unused `isTwoPhaseScriptAddress`
+* Deprecate `babbageTxScripts` and `refScripts`
 
 ## 1.4.5.0