+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version history for `cardano-ledger-api`
 
-## 1.6.0.1
+## 1.6.1.0
 
 *